'Сделал thunk для чекбокса, изменил пост запрос для нормального рендера добавил preloader, добавил confirm для удаления '

+++ src/Components/Task/TaskItem.tsx
@@ -2,7 +2,7 @@ import {Task} from "../Todo/TodoSlice.ts";
 import {FC} from "react";
 import {useDispatch} from "react-redux";
 import {AppDispatch} from "../../store.ts";
-import {deleteTodo,} from "../Todo/TodoThunks.ts";
+import {deleteTodo, updateTodoStatus,} from "../Todo/TodoThunks.ts";
 
 interface Props {
     task: Task;
@@ -12,13 +12,18 @@ const TaskItem: FC<Props> = ({task}) => {
     const dispatch: AppDispatch = useDispatch();
 
     const handleDelete = () => {
-        dispatch(deleteTodo(task.id));
+        if (window.confirm('Вы уверены?'))
+            dispatch(deleteTodo(task.id));
+    };
+
+    const handleChange = () => {
+        dispatch(updateTodoStatus({id: task.id, status: !task.status}));
     };
     return (
         <>
             <div className="card mb-3 container">
                 <div className="card-body">
-                    <input type='checkbox' checked={task.status}/>
+                    <input type='checkbox' checked={task.status} onChange={handleChange}/>
                     <p className='card-text'>{task.title}</p>
                 </div>
                 <div>
